Move work description list data creation to helper
In order to get a clearer overview of component
